new directories setup
