+++ client/src/authHandler.js
